Cargo fmt + move to modules
